fix: fix eslint warning and error
