+++ rs-react-app/src/pages/about/ui/about.module.css
@@ -35,7 +35,7 @@
 .logo {
   width: 3rem;
   height: 3rem;
-  background-image: url('../../../public/assets/img/rs.png');
+  background-image: url('/assets/img/rs.png');
   background-size: 100%;
   border-radius: 10%;
 }
