+++ rs-react-app/src/shared/context/page-context.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { PageContextType } from '../interfaces';
+
+export const PageContext = createContext<PageContextType>(null);
