+++ rs-react-app/src/shared/context/snackbar-context.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+export type SnackbarContextType = {
+  showSnackbar: (message: string, isError?: boolean) => void;
+};
+
+export const SnackbarContext = createContext<SnackbarContextType | undefined>(
+  undefined
+);
