+++ rs-react-app/src/shared/context/theme-context.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+import { ThemeContextType } from '../interfaces';
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined
+);
