+++ rs-react-app/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/main.tsx","./src/vite-env.d.ts","./src/__tests__/about-page.test.tsx","./src/__tests__/card-list.test.tsx","./src/__tests__/card.test.tsx","./src/__tests__/error-boundary.test.tsx","./src/__tests__/header.test.tsx","./src/__tests__/loading-component.test.tsx","./src/__tests__/main-page.test.tsx","./src/__tests__/not-found-page.test.tsx","./src/__tests__/page-context-provider.test.tsx","./src/__tests__/paginated-list.test.tsx","./src/__tests__/pagination.test.tsx","./src/__tests__/pokemon-selection-slice.test.tsx","./src/__tests__/product-detail.test.tsx","./src/__tests__/refresh-all-selected-button.test.tsx","./src/__tests__/refresh-pokemon-button.test.tsx","./src/__tests__/search-input.test.tsx","./src/__tests__/selection-panel.test.tsx","./src/__tests__/store.test.ts","./src/__tests__/theme-provider.test.tsx","./src/__tests__/use-local-storage.test.tsx","./src/app/app.tsx","./src/app/index.ts","./src/app/router.tsx","./src/app/store/index.ts","./src/app/store/root-reducer.ts","./src/app/store/store.ts","./src/features/index.ts","./src/features/pokemon-api/index.ts","./src/features/pokemon-api/pokemon-api.ts","./src/features/pokemon-selection/index.ts","./src/features/pokemon-selection/model/index.ts","./src/features/pokemon-selection/model/slice.ts","./src/features/pokemon-selection/ui/index.ts","./src/features/pokemon-selection/ui/refresh-all-selected-button.tsx","./src/features/pokemon-selection/ui/selection-panel.tsx","./src/pages/index.ts","./src/pages/about/index.ts","./src/pages/about/ui/about-page.tsx","./src/pages/about/ui/index.ts","./src/pages/main/index.ts","./src/pages/main/ui/index.ts","./src/pages/main/ui/main-page.tsx","./src/pages/main/ui/components/card-list.tsx","./src/pages/main/ui/components/card.tsx","./src/pages/main/ui/components/index.ts","./src/pages/main/ui/components/paginated-list.tsx","./src/pages/main/ui/components/pagination.tsx","./src/pages/main/ui/components/product-detail.tsx","./src/pages/main/ui/components/refetch-button.tsx","./src/pages/main/ui/components/refresh-pokemon-button.tsx","./src/pages/main/ui/components/search-input.tsx","./src/pages/not-found/index.ts","./src/pages/not-found/not-found-page.tsx","./src/shared/index.ts","./src/shared/context/index.ts","./src/shared/context/page-provider.tsx","./src/shared/context/snackbar-provider.tsx","./src/shared/context/theme-provider.tsx","./src/shared/hooks/index.ts","./src/shared/hooks/use-app-dispatch.ts","./src/shared/hooks/use-app-selector.ts","./src/shared/hooks/use-snackbar.ts","./src/shared/hooks/use-storage.ts","./src/shared/hooks/use-theme.ts","./src/shared/interfaces/index.ts","./src/shared/interfaces/interfaces.ts","./src/shared/ui/async-button.tsx","./src/shared/ui/error-boundary-navigate.tsx","./src/shared/ui/error-boundary.tsx","./src/shared/ui/header.tsx","./src/shared/ui/index.ts","./src/shared/ui/layout.tsx","./src/shared/ui/loading-component.tsx","./src/shared/ui/snackbar.tsx","./src/shared/ui/theme-toggle.tsx","./src/utils/index.ts","./src/utils/test-utils.tsx","./test/jest.setup.ts","./vite.config.ts"],"errors":true,"version":"5.8.3"}
